% File src/library/grid/man/grid-internal.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2013 R Core Team
% Distributed under GPL 2 or later

\name{grid-internal}
%% Part of the sensible export list but (currently?) documented as
%% internal (or waiting for documentation to be written):
%% <LIST>
%%\alias{makethreeinter_cpp}
\alias{expit}
\alias{hmc4sfa}
\alias{make.Z_EM}
\alias{make.Z_gibbs}
\alias{test.gamma.pois_EM}
\alias{test.gamma.pois_gibbs}
\alias{update_UDV_EM}
\alias{update_UDV_gibbs}
\title{Internal SFA Functions}
\description{
  Internal SFA functions.
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
