\name{SparseFactorAnalysis-package}
\alias{SparseFactorAnalysis-package}
\docType{package}
\title{
Sparse factor analysis for mixed binary and count data.
}
\description{
 Multi-dimensional scaling provides a means of uncovering a latent structure underlying observed data.  This package presents a means for scaling binary and count data, for example the votes and word counts for legislators.
}
\details{
\tabular{ll}{
Package: \tab SparseFactorAnalysis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-21\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Marc Ratkovic, In Song Kim, John Londregan, and Yuki Shiraito
Maintainer: Marc Ratkovic (ratkovic@princeton.edu)
}
\references{
In Song Kim, John Londregan, and Marc Ratkovic.  2015.  "Voting, Speechmaking, and the Dimensions of Conflict in the US Senate."  Working paper.
}
\keyword{ package }

